+++ lib/domain/initial_setting/initial_setting_2_page.dart
@@ -2,7 +2,7 @@ import 'package:Pilll/router/router.dart';
 import 'package:Pilll/store/initial_setting.dart';
 import 'package:Pilll/components/atoms/buttons.dart';
 import 'package:Pilll/components/atoms/color.dart';
-import 'package:Pilll/domain/initial_setting/initial_setting_3.dart';
+import 'package:Pilll/domain/initial_setting/initial_setting_3_page.dart';
 import 'package:Pilll/components/organisms/pill/pill_sheet.dart';
 import 'package:Pilll/components/atoms/font.dart';
 import 'package:Pilll/components/atoms/text_color.dart';
