+++ lib/domain/initial_setting/initial_setting_page.dart
@@ -1,11 +1,11 @@
 import 'package:Pilll/components/atoms/buttons.dart';
 import 'package:Pilll/components/atoms/color.dart';
-import 'package:Pilll/domain/initial_setting/initial_setting_1.dart';
+import 'package:Pilll/domain/initial_setting/initial_setting_1_page.dart';
 import 'package:Pilll/components/atoms/font.dart';
 import 'package:Pilll/components/atoms/text_color.dart';
 import 'package:flutter/material.dart';
 
-class InitialSetting extends StatelessWidget {
+class InitialSettingPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
